% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_brewer.R
\name{uoe_brewer}
\alias{uoe_brewer}
\title{University of Exeter palette generator}
\usage{
uoe_brewer(
  palette_name,
  n,
  type = c("discrete", "continuous"),
  direction = c(1, -1),
  override_order = FALSE
)
}
\arguments{
\item{n}{Number of desired colours. If number of requested colours is beyond the scope of the palette, colours are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically interpolate between colours.}

\item{direction}{Sets order of colours. Default palette is 1. If direction is -1, palette colour order is reversed}

\item{override_order}{Colours are picked from palette to maximise readability and aesthetics. This means that colours are not always selected in sequential order from the full palette. If override.order is set to TRUE, colours are selected in sequential order from the full palette instead. Default is FALSE.}

\item{palette}{Name of Palette.}
}
\value{
A vector of colours.
}
\description{
Color palettes for the University of Exeter 2022 branding. Code taken from MetBrewer.
}
