% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_colours.R
\name{uoe_colours}
\alias{uoe_colours}
\title{Complete list of colours.}
\usage{
uoe_colours(cols = NULL)
}
\arguments{
\item{cols}{character vector of colours to return.}
}
\value{
returns the hex codes from the University of Exeter colour palettes by their name.
}
\description{
Use \code{names(uoe_colours())} to return all possible colour names.
Use \code{\link{uoe_palettes}} to view palettes.
}
