% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_palettes.R
\docType{data}
\name{uoe_palettes}
\alias{uoe_palettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 1.
}
\usage{
uoe_palettes
}
\description{
Use \code{names(uoe_palettes)} to return all possible palette names. Current choices are:
\code{primary}, \code{secondary}, or \code{combined}.
Use \code{\link{uoe_brewer}} to construct palettes.
}
\keyword{datasets}
